Added Camera Calibration code
In this script, you need to have selected 6 points that have known measurements in one camera angle and another camera angle that those same six points can be seen again. Also the camera, you'll notice an array called object points which are the points in which I selected and their exact measurements.